refactor(download): remove agentweb-download module , Downloader replace implement
