+++ .idea/modules.xml
@@ -4,7 +4,6 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/AgentWeb.iml" filepath="$PROJECT_DIR$/AgentWeb.iml" />
       <module fileurl="file://$PROJECT_DIR$/agentweb-core/agentweb-core.iml" filepath="$PROJECT_DIR$/agentweb-core/agentweb-core.iml" />
-      <module fileurl="file://$PROJECT_DIR$/agentweb-download/agentweb-download.iml" filepath="$PROJECT_DIR$/agentweb-download/agentweb-download.iml" />
       <module fileurl="file://$PROJECT_DIR$/agentweb-filechooser/agentweb-filechooser.iml" filepath="$PROJECT_DIR$/agentweb-filechooser/agentweb-filechooser.iml" />
       <module fileurl="file://$PROJECT_DIR$/sample/sample.iml" filepath="$PROJECT_DIR$/sample/sample.iml" />
     </modules>
