+++ agentweb-core/build.gradle
@@ -38,6 +38,7 @@ dependencies {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
     testCompile 'junit:junit:4.12'
+    compile 'com.github.Justson:Downloader:v4.1.0'
     compileOnly "com.android.support:design:${SUPPORT_LIB_VERSION}"
     compileOnly "com.android.support:support-v4:${SUPPORT_LIB_VERSION}"
     compileOnly fileTree(include: ['*.jar'], dir: 'providedLibs')
