+++ agentweb-core/src/main/java/com/just/agentweb/AgentWebSettingsImpl.java
@@ -22,35 +22,21 @@ import android.webkit.WebView;
 
 
 /**
- * @since 1.0.0
  * @author cenxiaozhong
+ * @since 1.0.0
  */
 public class AgentWebSettingsImpl extends AbsAgentWebSettings {
-    private AgentWeb mAgentWeb;
+	private AgentWeb mAgentWeb;
 
-    @Override
-    protected void bindAgentWebSupport(AgentWeb agentWeb) {
-        this.mAgentWeb = agentWeb;
-    }
+	@Override
+	protected void bindAgentWebSupport(AgentWeb agentWeb) {
+		this.mAgentWeb = agentWeb;
+	}
 
-    @Override
-    public WebListenerManager setDownloader(WebView webView, DownloadListener downloadListener) {
-        Class<?> clazz = null;
-        Object mDefaultDownloadImpl$Extra = null;
-        try {
-            clazz = Class.forName("com.just.agentweb.download.DefaultDownloadImpl");
-            mDefaultDownloadImpl$Extra =
-                    clazz.getDeclaredMethod("create", Activity.class, WebView.class,
-                            Class.forName("com.just.agentweb.download.DownloadListener"),
-                            PermissionInterceptor.class)
-                            .invoke(mDefaultDownloadImpl$Extra, (Activity) webView.getContext()
-                                    , webView, null, mAgentWeb.getPermissionInterceptor());
-
-        } catch (Throwable ignore) {
-            if (LogUtils.isDebug()) {
-                ignore.printStackTrace();
-            }
-        }
-        return super.setDownloader(webView, mDefaultDownloadImpl$Extra == null ? downloadListener : (DownloadListener) mDefaultDownloadImpl$Extra);
-    }
+	@Override
+	public WebListenerManager setDownloader(WebView webView, DownloadListener downloadListener) {
+		return super.setDownloader(webView, downloadListener == null ?
+				DefaultDownloadImpl.create((Activity) webView.getContext(), webView, mAgentWeb.getPermissionInterceptor()) :
+				downloadListener);
+	}
 }
