--- agentweb-download/build.gradle
@@ -1,62 +1,0 @@
-apply plugin: 'com.novoda.bintray-release'
-apply plugin: 'com.android.library'
-apply plugin: 'maven'
-
-android {
-    compileSdkVersion COMPILE_SDK_VERSION.toInteger()
-    buildToolsVersion "${BUILD_TOOL_VERSION}"
-
-
-    defaultConfig {
-        minSdkVersion 9
-        targetSdkVersion TARGET_SDK_VERSION.toInteger()
-        versionCode 1
-        versionName "1.0"
-
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-
-    lintOptions {
-        abortOnError false
-    }
-    sourceSets{
-        main{
-            jniLibs.srcDir(['libs'])
-        }
-    }
-
-    repositories {
-        flatDir {
-            dirs 'libs'
-        }
-    }
-}
-
-dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    testCompile 'junit:junit:4.12'
-    compile 'com.github.Justson:Downloader:v4.0.3'
-    androidTestCompile('com.android.support.test.espresso:espresso-core:3.0.1', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
-    testCompile 'junit:junit:4.12'
-    provided "com.android.support:design:${SUPPORT_LIB_VERSION}"
-    provided "com.android.support:support-v4:${SUPPORT_LIB_VERSION}"
-    compile project(':agentweb-core')
-}
-publish {
-    userOrg = 'just'
-    groupId = 'com.just.agentweb'
-    artifactId = 'download'
-    publishVersion = '4.0.3-beta'
-    desc = 'very easy to build you web page'
-    website = "https://github.com/Justson/AgentWeb"
-}
