--- agentweb-download/src/main/AndroidManifest.xml
@@ -1,8 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.just.agentweb.download">
-
-    <application
-        >
-    </application>
-
-</manifest>
